// src/base/_/application/exception/dto/exception.enum.ts
enum ExceptionEnum {
  ALREADY_EXISTS = 'ALREADY_EXISTS',
  CONFLICT = 'CONFLICT',
  FORBIDDEN = 'FORBIDDEN',
  INVALID = 'INVALID',
  NOT_FOUND = 'NOT_FOUND',
  UNAUTHORIZED = 'UNAUTHORIZED',
  BAD_REQUEST = 'BAD_REQUEST',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
  OTHER = 'OTHER',
}

export default ExceptionEnum;